various changes to decrease interdependencies using #ifdefs; animated text and speech text ; party (unverified functionality); corrected 7.5 fingerprint; trade is now rejected; perhaps some more?
git-svn-id: svn://ivan.vucica.net/theoutcast@78 4617c47b-1226-0410-b705-8e5dc674c1d5

+++ theoutcast/gm_logo.cpp
@@ -96,11 +96,11 @@ void GM_Logo::Render() {
 		glEnable(GL_BLEND);
 		glBegin(GL_QUADS);
 		if (fadein) {
-		    SoundSetMusicVolume((1.-fadein) * 128);
+		    SoundSetMusicVolume((int)((1.-fadein) * 128));
 			glColor4f(0,0,0, fadein);
 		}
 		else {
-		    SoundSetMusicVolume((fadeout) * 128);
+		    SoundSetMusicVolume((int)((fadeout) * 128));
 			glColor4f(0,0,0,1. - fadeout);
 		}
 		glVertex2f(0, 0);
